server/requirements.txt
eventlet
python-socketio
cryptography
pyjwt